requirements.txt
docling
httpx>=0.27
tenacity>=8.2
pydantic>=2.7
loguru>=0.7
rapidfuzz>=3.9
python-dateutil
python-dotenv
tqdm
ujson
lxml
pillow
numpy
pandas
pypdfium2
piexif
# Optional local concept linkers (install only if needed)
# scispacy
# spacy>=3.5
# quickumls